Example project
Example project created as base for running unit tests. Individual pod files tests need to be tied to xcodeproj in order to run. To resolve that problem we have example project that can be also used as runtime. This limitations comes as we use more than one module (multiple podspec files)